requirements.txt
mcp>=1.13.0
watchdog[watchmedo]>=3.0.0
uvicorn>=0.24.0
fastapi>=0.104.0
httpx>=0.25.0
pyyaml>=6.0
pydantic>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
